#18 refactor: 서비스 클래스 위치 및 이름 변경
- RedisService와 MongoService 클래스의 위치를 infrastructure 패키지에서 health 패키지로 이동
- TestMongoService를 MongoService로 이름 변경
- 관련 import 문 수정
- 클래스 내부의 메소드 및 변수명 변경
